Add delay for reboot on QEMU for 24.04
Unlike 22.04, Ubuntu 24.04 doesn't appear to kill the SSH proxy
consistently (forcing a reconnect) on reboot.

This causes the QEMU provisioner to hang waiting for the connection,
which has reset but is still technically alive. Instead add a 10 second
delay (which is likely less than the time to reboot) so the SSH command
fails whilst we boot, and packer retries connecting correctly.